requirements.txt
Flask==1.0.2
gunicorn
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.4
boto==2.48.0
bokeh==0.12.16
urllib3==1.22
requests==2.18.4
tornado==5.0.2
six==1.11.0
pandas==0.23.0
packaging==17.1
numpy==1.14.5
python-dateutil==2.7.3
Jinja2==2.10
PyYAML==3.12 
MarkupSafe==0.18
Werkzeug>=0.14
itsdangerous==0.23
